Particle system complete, Wraith mechanics implemented, and more
- Renamed Rogue to Wraith
- Added WIP Quantum Flux shader
- Added Flare effect sprite
- Added Flare2 effect sprite
- Added Glow effect sprite
- Boss Shadow C sprite update
- Added Quantum UI sprites
- Added Quantum UI
- Zephyr Armor sprites updated
- Added Zephyr Veil
- Added Wraith variables to GlobalItem and ModPlayer
- InRadians and InDegrees accepts floats again
- Added Criadryn Spikes
- Updated Wraith Item base
- Updated Criadryn movement
- Updated Criadryn aiming, added prediction
- Primordial Caelus changes
- Added Moon Blast Particle
- Particle System autofinds texture for particle
- Particle System draw function works properly
- Particle System variables camel-cased
- Added scale parameter for particle spawning
- Updated CriadrynSpike AI
- Updated Moon Blast visuals
- Added Criadryn Spike Thrown projectile for Wraiths
- Added hook for Particle System to draw before Dusts, in ProvidenceHooks
- Made IsThereABoss util Tuple more readable
- Added AnimationFrame overload for Texture2D

+++ Projectiles/Cleric/StormfrontTempest.cs
@@ -46,7 +46,7 @@ namespace ProvidenceMod.Projectiles.Cleric
 			Player player = Main.player[projectile.owner];
 			if (projectile.timeLeft > 30)
 				projectile.ai[0] += (angleMod < 24f ? angleMod += 0.02f * (1f + angleMod) : angleMod);
-			Vector2 pos = new Vector2(100f, 0f).RotatedBy(projectile.ai[1]).RotatedBy( ((double) projectile.ai[0]).InRadians());
+			Vector2 pos = new Vector2(100f, 0f).RotatedBy(projectile.ai[1]).RotatedBy( projectile.ai[0].InRadians());
 			projectile.rotation = pos.ToRotation();
 			projectile.Center = new Vector2(player.Center.X, player.Center.Y) + pos;
 			if (projectile.timeLeft < 31)
